extract Accept Bid workflow
