+++ src/containers/flows/AcceptBid/constants.ts
@@ -1,0 +1,41 @@
+// Status Text
+export const CHECK_GAS_STATUS =
+	'Ensuring you have enough gas to approve zAuction...';
+export const ACCEPT_ZAUCTION_PROMPT = `Before you can accept a bid, your wallet needs to approve zAuction. You will only need to do this once. This will incur gas fees. 
+	\nPlease accept in your wallet...`;
+export const APPROVING_ZAUCTION =
+	'Approving zAuction. This may take up to x mins. Do not close this window or refresh your browser as this may incur additional gas fees.';
+export const ACCEPTING_BID =
+	'Accepting Bid... This may take up to x mins. Do not close this window or refresh your browser...';
+export const CHECKING_ZAUCTION_STATUS =
+	'Checking status of zAuction Approval...';
+export const AWAITING_APPROVAL = 'Waiting for approval from your wallet...';
+
+// Error Text
+export const NO_ERROR = '';
+export const LOW_BALANCE = "You don't have enough Ether to use as gas";
+export const FAIL_TO_CALCULATE_GAS =
+	'Failed to calculate gas costs. Please try again later.';
+export const APPROVAL_REJECTED = 'Approval rejected by wallet';
+export const FAILED_TRANSACTION = 'Failed to submit transaction.';
+
+// Step Prompt Text
+export const ZAUCTION_PROMPT =
+	'Before you can accept a bid, your wallet needs to approve zAuction. You will only need to do this once. This will incur gas fees.';
+export const CONFIRM_BID_AMOUNT = 'Are you sure you want to accept a bid of';
+export const SUCCESS_CONFIRMATION =
+	'Success! Bid accepted and ownership transferred';
+
+// Button Text
+export const CANCEL_BTN = 'Cancel';
+export const CONTINUE_BTN = 'Continue';
+export const FINISH_BTN = 'Finish';
+
+// Labels
+export const WILD_CURRENCY_CODE = 'WILD';
+export const USD_CURRENCY_CODE = 'USD';
+export const LOADING_BIDS = 'Loading bids...';
+export const CREATOR_LABEL = 'Creator';
+export const HIGHEST_BID_LABEL = 'Highest Bid';
+export const SELECTED_BID_LABEL = 'Bid You’re Accepting';
+export const ROOT = '0://';
